+++ src/main/java/com/bous/calisthenics_app/repository/WorkoutLogRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface WorkoutLogRepository extends JpaRepository<WorkoutLog, Long> {
-    List<WorkoutLog> findByUserId(int userId);
-    List<WorkoutLog> findByUserIdAndExerciseId(int userId, Long exerciseId);
+    List<WorkoutLog> findByUserId(Long userId);
+    List<WorkoutLog> findByUserIdAndExerciseId(Long userId, Long exerciseId);
 }
